+++ .gitignore
@@ -1,5 +1,6 @@
 .idea
 .vscode
+tsconfig.tsbuildinfo
 
 # dependencies
 /node_modules
